refactor: moved services into separate folder
