+++ src/app/spell-lookup/services/spell-lookup-http-error-handler.service.ts
