+++ src/app/spell-lookup/services/spell-lookup.service.spec.ts
